Rename buggy file to fix-phone-num.py
